Borrando archivos y carpetas no ordenadas

+++ unidad 1/Clase6Victor/back/api/models/persona.py
@@ -3,8 +3,8 @@ from django.db import models
 class Persona(models.Model):
 	nombre 				= models.TextField() #'CharField' Campo de texto
 	apellido 				= models.TextField()
-	estatura 		= models.DecimalField(max_digits=4, decimal_places=2)
-	peso 		= models.DecimalField(max_digits=4, decimal_places=2)
+	estatura 		= models.IntegerField()
+	peso 		= models.IntegerField()
 	
 	def __str__(self):
 		return self.nombre
